feat(db): implement database restructuring v2 — content→main_entry, content_details merge
Implements D-198 (plan v2):
- Rename content → main_entry (all .sq files + Kotlin callers)
- Merge anilist_detail + extension_detail + other_source_detail + anime_metadata_cache
  → content_details (26 cols, data_*/ext_* prefixes, nullable axes)
- Drop app_metadata (absorbed into app_settings)
- Keep data_source + system separate (per R-2)
- Keep extension_repo_id (user directive)
- Keep display_source as single UX column (axis-level values)
- Fix missing FKs on watch_progress + notification_sent
- Fix episode_number type (INTEGER→REAL, Long→Double in Kotlin)
- Add ContentDetails + DataSourceExtras + ExtensionExtras data classes
- Add clearDataSourceAxis/clearExtensionAxis (fixes orphan-row unlink bug)
- Add unlinkExtension method to ContentResolver
- Add updateMainEntryTitle query
- Drop dead queries, methods, indexes
- Update DatabaseDriverFactory for schema rebuild

40 files changed. CI build will verify compilation.

+++ ANI-KUTA/APP/ani-kuta/core/notifications/src/main/java/com/confused/anikuta/core/notifications/NotificationConfigStore.kt
@@ -74,7 +74,7 @@ class NotificationConfigStore(
     /** Check if a notification was already sent (dedup). */
     suspend fun wasNotified(
         mainId: String,
-        episodeNumber: Long,
+        episodeNumber: Double,
         audioVariant: String,
         triggerType: String,
     ): Boolean = withContext(Dispatchers.IO) {
@@ -86,7 +86,7 @@ class NotificationConfigStore(
     /** Record a sent notification (for dedup). */
     suspend fun recordSent(
         mainId: String,
-        episodeNumber: Long,
+        episodeNumber: Double,
         audioVariant: String,
         triggerType: String,
     ) = withContext(Dispatchers.IO) {
